refactor(agents): the degrade concept gets its own file
runner.go crossed the 500-line ceiling. The closed reason vocabulary and
the two degrade constructors are one concept with one rule behind it —
what a person may read and what only an operator may — so they move
together, which also puts that rule where somebody adding a reason finds
it.
